popup을 통해 usermenu생성
