+++ shared/components/header/HomeIcon.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 export default function HomeIcon() {
   return (
     <Link href={'/'}>
-      <Image src="/logo.png" alt="homelogo" width={50} height={50} />
+      <Image src="/logo.png" alt="homelogo" width={60} height={60} />
     </Link>
   );
 }
