+++ shared/components/header/auth/index.tsx
@@ -1,17 +1,10 @@
 import { auth } from '@/auth';
-import { SignIn } from './AuthComponent';
-import Image from 'next/image';
+import { SignIn } from './SignIn';
+import UserMenu from './UserMenu';
 
 export default async function AuthButton() {
   const session = await auth();
   if (!session?.user) return <SignIn />;
 
-  return (
-    <Image
-      src={session.user.image!}
-      alt="내구글이미지"
-      width={50}
-      height={50}
-    />
-  );
+  return <UserMenu image={session.user.image!} />;
 }
