+++ shared/components/header/header.module.css
@@ -1,12 +1,12 @@
 .header {
   display: flex;
   justify-content: space-between;
+  position: sticky;
   align-items: center;
   padding: 10px 24px;
-  position: sticky;
   top: 0;
   background-color: white;
-  z-index: 10;
+  z-index: 800;
   border-bottom: 1px solid gray;
 }
 
@@ -26,4 +26,43 @@
   display: flex;
   flex-direction: row;
   gap: 1rem;
+}
+
+.user_menu {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.pop_over {
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  bottom: 0;
+  right: 0;
+  padding: 8px 4px;
+  transform: translate(-10px, 100%);
+  width: 192px;
+  background-color: #323333;
+  border: 1px solid rgb(2, 22, 49);
+  border-radius: 8px;
+  align-items: center;
+  gap: 8px;
+}
+
+.pop_button {
+  display: block;
+  border: none;
+  background: none;
+  cursor: pointer;
+  color: white;
+  text-decoration: none;
+  width: 100%;
+  text-align: center;
+  padding: 4px;
+}
+
+.pop_button:hover {
+  color: rgb(191, 198, 136);
 }
