+++ shared/components/portal/portal.module.css
@@ -5,7 +5,7 @@
   bottom: 0;
   left: 0;
   background-color: rgba(0, 0, 0, 0.7);
-  z-index: 910;
+  z-index: 900;
 }
 
 .modal {
