+++ shared/styles/globals.css
@@ -14,3 +14,10 @@ body {
   height: 100vh;
   overflow: hidden;
 }
+
+.no_style_btn {
+  padding: 0;
+  border: none;
+  background: none;
+  cursor: pointer;
+}
